Create structure to allow for sorting by films or cinema
