+++ src/assets/scss/_hero.scss
@@ -51,7 +51,7 @@
   position: relative;
   font-size: 3.25rem;
   white-space: pre-wrap;
-  max-width: 37rem;
+  max-width: 32rem;
   opacity: 0;
   animation: slide-up-far 600ms ease forwards;
   -webkit-animation: slide-up-far 600ms ease forwards;
@@ -63,9 +63,9 @@
 
 .hero-subtitle {
   position: relative;
-  font-family: "Oswald", sans-serif;
-  font-size: 1.6rem;
-  max-width: 41rem;
+  font-family: "Kanit", sans-serif;
+  font-size: 1.5rem;
+  max-width: 40rem;
   opacity: 0;
   animation: slide-up-far 500ms ease 400ms forwards;
 	-webkit-animation: slide-up-far 500ms ease 400ms forwards;
