+++ src/assets/scss/_variables.scss
@@ -1,5 +1,5 @@
 @import url('https://fonts.googleapis.com/css2?family=Open+Sans:ital,wght@0,300..800;1,300..800&display=swap');
-@import url('https://fonts.googleapis.com/css2?family=Oswald:wght@200..700&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Kanit:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap');
 
 @font-face {
   font-family: "BebasNeue";
@@ -11,7 +11,7 @@
 // $color-primary-highlight: #ff98a4;
 // $color-primary: #8c4ade;
 $color-primary: #ba98ff;
-$color-primary-highlight: #f8556e;
+$color-primary-highlight: #ff5569;
 
 // $color-secondary: #e2be2f;
 $color-secondary: #67dcfc;
